[API change] introduce RenderPluginModel
RenderPluginModel contains the business logic for managing plugins via the GUI, which was previously duplicated in MarblePart and QtMarbleConfigDialog.
In addition, RenderPluginModel allows for introducing new kinds of plugins w/o having to change code in the GUI.

The API change comprises the follwing changes in RenderPlugin:
(1) the ItemDataRole enum is moved to RenderPluginModel.
(2) item(), applyItemState() and retrieveItemState are now private because they are helpers of the RenderPluginModel (and are therefore considered to be implementation details)
Note that item() in combination with ItemDataRole used to provide only redundant information, which can be retrieved from RenderPlugin directly.
For instance, a RenderPlugin can be tested if it has a configuration dialog by dynamic_cast'ing it to Marble::DialogConfigurationInterface and checking whether it returns non-zero.

CCMAIL kde-bindings@kde.org