New dgml tag <vectortile>, with its parser and new GeoScenes
For GSoC 2012 vector tile rendering, a new tag <vectortile> has been
created for the dgml format and the handler for it in the dgml parser.
In te begining it was thought to call it <vectorTile> as some composed
words tags have also camel casing. But through the code this tile is
compared with the backend=vectortile .tolower() tag and it didn't match
vectortile != vectorTile. So for consistency the tag will be
<vectortile>.

Now that there are two possible tiled layers (Image tile layers with
<texture> tag and Vector tile layers with <vectortile> tag)
GeoSceneTexture has been turned to GeoSceneTextureTile which extends a
GeoSceneTiled abstract class. Also a GeoSceneVectorTile class has been
created.

GeoSceneTiled stores all the common data for layers containing tiles
and then GeoSceneTextureTile and GeoSceneVectorTile can manage specific
data related to them. Currently booth have the same dgml structure so
this two classes have no specific attribute but they are needed
afterwards for MarbleMap to know if its a VectorTile or a ImageTile
layer.

REVIEW: 106009