SatellitesPlugin: Add plugin activation tab page in config dialog
This adds a new tab page to the config dialog that is shown if the plugin
is not active. All other tab pages are hidden in this case. The user can
click a button to activate the plugin. If he or she do this, the plugin
configuration tab pages show up. This avoids confusion with empty catalog
lists. These lists are only available if the plugin is activated.