aprs: use external QextSerialPort for TTY reading
Instead of embedding an (old) copy of the QextSerialPort library, find for an external one;
only if found enable the reading from TTY, which is otherwise disabled (leaving its configuration tab disabled).

The drop of the internal QextSerialPort should also fix all the portability issues,
since the plugin itself does not use any OS-dependent API, and it is then reenabled unconditionally.
Hence,
BUG: 241125
should now be fixed, and
CCBUGS: 237931, 242039
should not happen anymore.

REVIEW: 107536