Fix unwanted scrolling issue.
When using the scrollbars with the mouse (not wheel) and moving out
of the dialog while still pressing the mouse, then releasing the
mouse and coming back to the dialog, scrolling was continued even
when the mouse was no longer on top of the scrollbars, leading to fast
and annoying unwanted scrolling. Now the mouse release that is needed
to stop such things is forwarded to the dialog after a mouse press
even if the dialog area was left already.