quick fix to make Marble's UI fit on netbook screens again
Marble's UI should fit on a netbook screen (minus e.g. the Plasma panel), even when Marble's status bar is visible.
To help achieve this goal, the following measures are applied to widgets that get wrapped in QDockWidgets:
* set minimum sizes of resizable widgets to zero (although not always respected)
* set margins to zero

Although this patch helps to preserve vertical space, more needs to be done for a real fix.
For instance, the RoutingWidget still consumes too much vertical space when via points are added, which has been a problem on Maemo for a while already.