Annotate: a plugin to draw polygons and placemarks on a map
The revival of osmannotate plugin which never got to release state.
The plugin layer handles SceneGraphicsItems for mouse event actions,
and maintains a document which is rendered with classical layers.
Still lots of work needed for production delivery, but already quite
functionnal:
- Adds/Removes Point Placemarks, Polygons
- Import/Export to kml file
- Reshape Polygons by moving points
- Placemark "edit bubble" which sets name and description

Needs work:
edit bubble is weak with painting and refreshing

REVIEW: 108618