fix part of coordinate grid disappearing due to rounding errors in conversion rad <-> deg when pole is visible
When compiling with clang and running on an Atom N270 (32 bit), the coordinate grid disappeared.
This happened because the comparison lon*RAD2DEG == 180 where lon == M_PI failed.
I detected that printing lon*RAD2DEG-180 resulted in non-zero output.
Strangely, I could not reproduce this when compiling with clang and running on 64 bit.

I think that the compiler shouldn't be blamed since IMO we shouldn't rely on the compiler doing the "correct" rounding for us.