remove quitZoom-to-quitRadius settings conversion code
Since commit 600cd553abfb2f212effbac1e2c07392e3b3cd49 the radius rather than the zoom value is stored in the settings.
This patch removes code which converted the "quitZoom" value into a radius value for maintainance reasons.
So if a user didn't start up Marble for half a year, the "quitZoom" settings value will be lost and a default radius will be used instead.