Fix compilation after API change in a411529
At least there should be a deprecation period for the removed methods.
Imho the removed drawLine method in a411529 is a convenience method
that is nice to have even. After all the library API should be easy to
explore and the implementation of the method is easy enough not to
introduce a maintenance problem.
CCMAIL: bbeschow@cs.tu-berlin.de