Add i18n context for satellite groups
This helps translators and also prevents a wrong translation of "Other"
or "Education" being used. Without additional context for example the
translation of "Education" (the satellite group) is taken from the
"marble" catalog (where "Education" means a map legend group) and not
from the correct "marble_qt" catalog. This is because "marble" is the
primary catalog (at least for the KDE GUI) and that is searched first.

REVIEW: 111753