Adjust test data to new writing behavior
We don't write tags with default values and do not guarantee that
the order of elements is preserved (both sane). GeoWriter test is
not aware of that, and the easiest way to fix it is to provide data
that fits (does not have tags with default values, input order matches
output order).