generate translation files for the Qt version automatically, if any
If translation files in .po format are present in ${src}/data/lang, .qm files will be generated automatically in the build directory, ready for installation during "make install".
The .po files can be downloaded by running the script "download-translations.bash" which also resides in ${src}/data/lang.

The motivation behind this change is to simplify packaging for the Maemo5 version where the build system had to be patched for installation of translations.

REVIEW: 114051