Hide ownCloud sync configuration.
Neither the ownCloud integration in Marble nor the Marble ownCloud app
have release quality yet. They are hidden now in the stable 1.7 branch
(KDE/4.12). If the remaining bugs are fixed it can be enabled by
default again even in the 1.7.x series.

There's a config setting to override this:
CloudSync/showCloudSyncSettings=true in the KDE or Qt config file,
respectively.