Hide the Tour panel from the Panels menu
The Tour Widget is not yet ready for prime time. It's not shown in
the Panels menu now, though it's still possible to activate it via
a right click on the menu area. Curious people reading commit messages
hence can still enable it, and we'll introduce it in full glory with
the next release.