Use a local path determined by Qt API as storage location in windows
While i cannot reproduce bug #337872 I suspect that the calls to
Windows API to determine a local storage are related. Since Qt has API
for it let's just use that instead.

At the same time the storage location changes from a roaming profile
location to a computer local location. If Marble stores several GB
in a roaming profile (happens easily for e.g. heavy OSM users), the
Windows login time can increase significantly when the roaming profile
is synced. Using a local location avoids that (at the cost of having
to install additional map themes on each computer in use separately).
If that breaks your use case, please open a bug report and I'll
introduce a config option to enable storing in roaming profiles again.

Old data in a roaming profile, if any, is migrated to the new location
on Marble startup if the user chooses so (a dialog asks about it once on
startup).

CCBUG: 337872