Change toolbuttons for Tour Ui.
After playing tour is stops, play and record buttons are enabled,
stop button is disabled, value of progress bar is set to 0.
If playlist doesn't contains "FlyTo", "Wait", "TourControl" play\record buttons are disabled.
Disable editing duration of first element of mainTrack if it's FlyTo and use 0 as it's duration.
Use QPersistentModelIndex where it's necessary.
REVIEW: 121751