Fixes bug related with zone labelling.
The previous code misbehaved when calculating longitude lines labels. Passing the longitude as a qreal to lonToString make that some lines get their zone label wrong (e.g., when passing -114, the function received -114.00000001 and returned the zone associated to [-120,-114[, not the one associated to [-114,-108[; i.e., it returned 11 instead of 12. The current code avoids this problem passing the longitude increased in 3 degrees (i.e., it does not pass the limit of the zone interval but the middle of it)