Improves the solution of the zone labelling bug. Fixes UTM grid. Fixes easting/northing calculation in the poles.
The solution implemented in a18adcfd4d2e7d754b1c030f620b1cd0f92ae245 is moved from renderLongitudeLines() in GraticulePlugin.cpp to the origin of the problem: lonLatToZone() function in GeoDataCoordinates.cpp. The zone labelling bug is no longer found (now, all the fuzzy values such as -114.00000001 are rounded to the nearest integer). Hence, lonLatToZone() function is now completely transparent to the exterior. The calls done from the outside, as from GraticulePlugin.cpp, do not need to shift its calculations to get the correct zone, even with the zone borders, where the bug was found.

The UTM grid showed the zone 31X shifted to the right, while it starts in its standard longitude. This commit fixes this bug with changes in renderUtmExceptions() in GraticulePlugin.cpp

The previous code misbehaved when calculating easting and northing in locations above 84N and below 80S. This commit fixes this bug with changes in lonLatToEasting() and lonLatToNorthing() in GeoDataCoordinates.cpp