Prefer iconPath to QImage instances in GeoDataIconStyle
Prefers passing a file path to instantiate GeoDataIconStyle instead of
passing a QImage directly. This leads to slightly faster startup times
(the icons of the static default styles are now lazy-loaded) and allows
querying the iconPath in more cases. The latter is beneficial for QML
images. Icon paths are also exported in MarblePlacemarkModel.

REVIEW: 124579