add missing main.moc include for plasmarunner and use cmake settings
why exactly the build works for anyone without the moc is beyond me, the
include however is supposedly necessary as documented at [1]. otherwise
the factory (object created by the macro) metaobject code is not resent
and the factory linking fails on the missing vtable. this is reproducible
on cmake 3.2.2, gcc5 and frameworks 5.13.

additionally include the KDECMakeSettings from ECM as to enable
current_binary_directory includes by default (which is necessary to
find main.moc)

[1] https://techbase.kde.org/Development/Tutorials/Plasma4/AbstractRunner#The_Plugin_Declaration_Macro