Routing layer must be informed of internal view context changes
Fixes a regression from input handler refactoring. Leads to a huge
performance improvement when panning/zooming the map and a route
with lots of points is shown.

A similar regression was fixed in cec5d95

This one was introduced in e6e217dd8fd90e31365c8ade42216b724e3438b4
and this commit pretty much reverts that commit.