Fix crash on exit
Member variables are destroyed in reverse order of appearance. When
the tile cache is destructed, the shared pointers of its documents
can reach zero and will attempt to remove themselves from the tree
model. While doing so the garbage queue and the list of pending
documents is checked, so they need to be intact at this point. This
is only the case if the tile cache appears after the garbage queue
and the list of pending documents in the member variables list.