Faster completion when there cannot be any results
When a completion has no results, a longer query starting with the same
string as the failing one cannot have any results (as long as the model
has not changed under the hood). Do not search in this case.

This speeds up the common case when the user keeps on typing for a
search string when the completion runs out of results, or hits backspace
repeatedly in the same case.