Use Qt5::* notation with target_link_libraries, & PUBLIC/PRIVATE interfaces
Summary:
Qt5::* notation is the usual way to refer to the imported cmake targets
representing the Qt5 modules/libs.

PUBLIC & PRIVATE link interfaces help to speed up compilation & linking,
because libs listed in PRIVATE will stay implementation detail of the lib
and not result in consumers of the lib also having those libs in the link
list. And it also reduces the amount of includes used, so less dirs to
search headers in.
To allow SailfishOS 2.0 builds, which has older cmake, still use
LINK_PUBLIC/LINK_PRIVATE for now

Reviewers: rahn, nienhueser

Reviewed By: nienhueser

Differential Revision: https://phabricator.kde.org/D1438