Allow plugins to be blacklisted and whitelisted
If you want to avoid that the plugin libFoo.so (Foo.dll on Windows) is
loaded at all, call Marble::PluginManager::blacklistPlugin("Foo").

If you want to avoid that any plugin other than libBar.so (Bar.dll
on Windows) is loaded at all, call
Marble::PluginManager::whitelistPlugin("Bar").

Each can be called multiple times with different plugins to blacklist
or whitelist several plugins. Plugins that are both blacklisted and
whitelisted are considered blacklisted.

Make sure to call the methods early enough before anything else
already used the PluginManager to load plugins. Take care to pass the
filename (without library prefix and file suffix) of the plugin, not
its nameId(). They can be identical, but often that is not the case.