return a static const default style instance instead of asking a static instance of StyleBuilder
GeoDataFeature::style() isn't used for rendering any more. Instead, the renderer asks the StyleBuilder to return a style.

Since there are many places in the code relying on style() returning a non-null pointer, we can't simply return a null pointer. Therefore, we return a dummy style object as a temporary solution.