Fix default crosshair having an offset
Summary:
Code assumed all images used for the crosshair have a size of 21x21.
The default image though is a PNG of a pre-rendered crosshair with a shadow,
which thus is not of that size.
That pre-rendered image, crosshairs-darkened.png, also had a size of 32x32,
while the crosshair in it had a pixel width of 1, so it's center would
not be pixel aligned with the cross lines. Image now resized to have
crosshair in the middle pixels.

Test Plan: Crosshair now is centered on middle, tested e.g. with 0, 0 lonlat.

Reviewers: nienhueser, #marble, rahn

Reviewed By: nienhueser, #marble, rahn

Differential Revision: https://phabricator.kde.org/D2138