Create placemarks with point geometry from non-trivial way nodes
Previously all nodes of the OSM file were examined (including those
part of ways and polygons) and completely discarded if their tags
did not result in a visual category or the visual category was a
highway one.

Now those nodes which are part of ways and polygons are removed early
in case they are trivial (have just a coordinate and an id). This is
fine because them being part of a way of polygon ensures all
information is still available. Otherwise they are kept.

This ensures that stale nodes (trivial but not referenced) are not
discarded anymore. Parsing should be slightly faster also given
that trivial nodes which are part of ways are not tried to be
converted to a placemark anymore.