AzimuthalProjection: Avoid duplicated globeHidesPoint calculation in iteration
Summary:
In the first recursive call there is no need to recalculate whether the
currentCoords parameter of AzimuthalProjectionPrivate::findHorizon(...)
is hidden by the globe, because that information is already known on the
previous level.
Splitting the findHorizon() method into a "interface" method and
a "worker" method for the recursion, where the existing information
is passed on, allows to avoid that duplicated calculation.

Reviewers: rahn, #marble

Reviewed By: rahn, #marble

Differential Revision: https://phabricator.kde.org/D2408