Improve node only vs node or ring vs node or area category styles
Many osm key/value combinations can be used on nodes (placemarks)
but also on ways or relations (linestring, linearring, polygon).
We map them to the same visual category, but the styling needs
to be different for ring and polygon features. This commit improves
the rendering by setting the outline vs filled rendering properties
according to how nodes are most often tagged.

This is a follow-up to 8ec3266 which avoided to fill features like
towns, but at the same time broke the rendering of e.g. parking places.
In the long run we need separate styles for placemark vs string vs
ring vs polygon features, ideally read from an external style
definition.