Port DeclarativeDataPlugin from QtScript to QJson* -> no more Qt5Script dep
With QJson support in Qt5Core, the dependency on Qt5Script can be avoided
(besides Qt5Script being deprecated).

This ports the last active code to use QJson* classes instead of the
QScript* ones. The port of DeclarativeDataPlugin is done blindly though,
as the code cannot be tested currently, because the "datalayers" property
of the MarbleQuickItem class was dropped in the Qt5 port.