Add batch rendering visualization for the desktop: This debug mode can be enabled via the "B" key stroke.
Ideally while rendering items the painter should perform a
minimal amount of Pen/Brush state changes. This goal is
reached when items with the same pen/brush assigned get
rendered in a single batch.

This visualization mode changes the color of the pen and brush
whenever a new different pen/brush is assigned.
As a result it's possible to observe

- different sub chunks of roads which should be merged into a single item
- objects that should be rendered in the same batch should have the
  same color for a given frame.

TODO:
- Add this feature to Marble Maps
- add defines to the debug methods to prevent compilation into
  the stable release.