More efficient street rendering:
Previously in order to draw a street the same street geometry was
fully reprojected and drawn up to three times: For the
Outline, Inline and Label.
This commit introduces means in ClipPainter and GeoPainter to
precalculate the projected geometry of a GeoDataLineString.
The result is used to render the outline, inline and to
draw the labels without doing a superfluous "transparent"
redrawing of the lineString (as it was previously done).
This also saves several pen-changes.

On a busy Level 11 area this commit improves performance
by about 20% (measured on a desktop machine).