Provide separate styles for buildings with shops etc
Features tagged as building=* and another visual tag get a style that
is a mixture of the building style (poly, line) and the visual tag
style (icon, label). Upon rendering the style is chosen greedily
based on the first visual tag found, if any. So plain buildings get
the style as before, and those that have e.g. shop=bakery get a
bakery icon on top. If there are more than one possible visual tags,
one of them is chosen at random.