Add a GeoDataRelation class, improve hiking route support
Similar to GeoDataContainer and GeoDataMultiGeometry. In contrast to
container types, placemarks are not owned, so sharing is possible. In
contrast to multi geometry, placemark meta data is preserved.

This class is needed to group semantically related placemarks without
duplicating them, useful e.g. to store hiking, ski, public transport
and similar routes and their alternatives.