Support translation catalogs added in KDE Application release tarball
Summary:
Starting with KDE Applications 17.04 the po files for the translations
are now added to the tarball of the sources they belong to, instead of
being released separately.
The release script also makes the assumption that all KDE Applications
use CMake, ECM and KI18n and simply inject the lines
ecm_install_po_files_as_qm(poqm)
ki18n_install(po)
to the end of the toplevel CMakeLists.txt.

Which does not work for Marble. To solve this, the release script
creating the KDE Applications tarballs gives Marble some special handling.

Additionally Marble's buildsystem needs to handle the case of Qt-only
building, where both macros are not available.

This patch solves this for now  by using a custom macro for the po->qm files,
and (in a temporary hacky way) calling ki18n_install in a subdir
where the macro is known due to linking against KI18n and which will be
skipt in qt-only builds.

For that this patch moves the existing code for handling the
po->qm translations for the bundle packaging of marble-maps into the
global MarbleMacros.cmake file and adapts it, so it can be reused also
for the release tarball po->qm files.

Test Plan:
Without po and poqm file build works as before.
Copied the po and poqm files from the released tarballs into the
toplevel dir, qm & mo files were created and installed as they should.

Custom bundle_translations target also works as before.

Reviewers: rahn, nienhueser, ltoscano, aacid

Reviewed By: nienhueser

Differential Revision: https://phabricator.kde.org/D5648