Introduced rotation the globe by right mouse button.
Summary: Also applied kinetic spinning like it is with changing position.

Test Plan:
to-check: open a map and push the right button of mouse, the globe is going to be rotated by moving mouse by x and y axis.
Also if you unclick the button, kinetic spinning will be enabled like it is when you use left mouse button.

Reviewers: #marble, nienhueser

Reviewed By: #marble, nienhueser

Subscribers: nienhueser, #marble

Tags: #marble

Differential Revision: https://phabricator.kde.org/D5650