Automatic download of .osm files in the "Edit Maps" Panel
I have added an icon in the Edits Maps Panel for downloading .osm file
for the current viewport(visible region)

    I have created a Dialog which opens on clicking the download icon(in
Edits Map Panel) and then in the Dialog on clicking download it tries to
download the .osm file from server.
    In case of network error or too large visible region an error
message will be shown telling to zoom in .
    It shows a Busy indicator progress bar since server does not support
reporting the size of the file.
    The .osm file is saved in a temporary directory
    If the download is successful then the file is opened automatically
using loadAnnotationfile(const QString& )

Differential Revision: https://phabricator.kde.org/D836