Make the bookmarks dialog responsive to touch everywhere in Marble Maps.
When we tap on the Search field in the app, if nothing is written in that field
then the bookmarks dialog shows up. If the number of bookmarks is <= 6, then the
far-lower part of the dialog does nothing on being tapped. Made that portion
choose the last item from the bookmarks list in that case.