Only process onClosing events on Android
Summary:
When Marble Maps are used on Android, it makes sense to modify "back" button behavior to avoid accidental app exit, simplify page stack navigation etc.
On Plasma Mobile (and Desktop, in convergent situations) we, however, want the application to be closed when a user swipes it away from their task manager.

Test Plan:
Verify that on Android the behavior has not changed.
Verify that on Plasma closing the window is enough to close the application.

Reviewers: #marble, mart

Reviewed By: mart

Subscribers: marble-devel, kde-edu, #marble

Tags: #marble, #kde_edu

Differential Revision: https://phabricator.kde.org/D18869