Rework Android APK generation
Summary: This reworks the APK generation to be more aligned to what the binary factory Android builder expects. Most of the heavy lifting is done there for us, so we don't need special code for it here.

Test Plan: Builds in local clone of the binary factory container

Reviewers: #marble, sanjibanb

Reviewed By: sanjibanb

Subscribers: apol, marble-devel, kde-edu

Tags: #marble, #kde_edu

Differential Revision: https://phabricator.kde.org/D24642