Fix precision loss when passing coordinates to the clipper library
This intends to convert coordinates to 100 nanodegree integer values
(the OSM base resolution), however it misses the fact that the input
is provided in radians, not degrees. This means that we are effectively
reducing the resolution by about 60x. This is enough to introduce visual
glitches up to completely destroying sub-meter structures.