Actually generate o5m tiles in the Tirex backend
This is mostly copy/pasting of the relevant code from cachetiles.cpp,
but this now actually allows to request on-demand created o5m tiles
via mod_tile, assuming the setup this code expects is all there.

This did expose a limitation of mod_tile, which assumes file extensions
to only contain lower-case letters, and thus chokes on our use of "o5m".
This can be worked around though by a simple rewrite rule.