Enable TileDirectory to load input data from an OSMX file
This is considerably more efficient to read and to update compared to the
manually maintained o5m pyramid used so far, at the cost of a lot of disk
space.

The OSMX usage is extremely primitive so far, relying solely on its command
line tool. We could gain a bit more performance still by using OSMX API
directly and avoiding the PBF file indirection, but that is in the ~1%
range, and technically not entirely straightforward due to how OSMX is
built and its bundled dependencies.

This code will be used by the upcoming Tirex backend for on-demand tile
generation, existing code is only touched for removing a small bit of
dead code.