Remove broken caching of area computation results
The problem here is that this code uses the address of the argument as
cache key. The argument can however be a temporary, meaning we can get
cache key collisions and thus wrong results. This shows in tiles
occasionally missing a random subset of polygons, typically in the zoom
range 11-15, where we apply area size filters.

The solution is fortunately easy: Remove the caching of the area
computation result entirely. We can do this as this neither contributes
to the overall cost in a measurable way, nor do we seem to compute the
area of the same polygon multiple times anyway.