Redo the communication code with the Tirex server
This is using a slightly weird setup with a 1:N UDP socket connection
for distributing work to multiple render processes. Qt's UDP code doesn't
really get along with that, and over time degenerates to only one of the
backends still being able to communicate.

To address this, bypass Qt entirely for the communication and do this
manually. That reduces portability, but that doesn't matter given that
Tirex and mod_tile aren't more portable either.