Make the input data load size more flexible
For this we need to split a meta tile into a number of sub-blocks and
traverse them in the right order to avoid loading input data multiple
times.

This allows us to tweak the load size for z13 in a way that is good enough
for live generation as well.