Work around MARBLE_ADD_WRITER not working in static builds
This relies on being executed during plugin loading, but in a static build
it's either executed too early when the called code isn't ready yet (which
crashes), or not at all.

This can probably be done nicer by refactoring this to avoid relying on
static construction, but given this is only actually used in one single
place handling this manually for the Tirex backend is good enough.