Don't discard any OSM nodes with tags during loading
This affects a very small amount of nodes that are not currently rendered
by Marble (e.g. various types of free-standing ticket vending machines).
With this change those are also included in the generated o5m tiles, which
matters for other users than Marble (and presumably Marble would want to
show those properly eventually too).

This also require a minor change to Marble's style selection, as unknown
nodes were shown with the default style (a yellow dot) so far, which is
undesired I assume (but made it very easy to verify this is only a small
amount in total).