Don't discard non-visual multi-polygon parts with tags during OSM loading
Those are extremely rare, and obviously have no visual impact in Marble,
but they do contain valuable information for other users of the vector
tiles.

An example are railway=platform_edge segments of railway platforms (and
that's actually the only case affected by this I have found so far), and
that's about 1.5k elements for all of Germany, so not an amount that would
matter in the big picture.