Skip creation of temporary files for zipped KML (KMZ)
Instead of extracting all files in the ZIP archive extract only the
KML file into a QByteArray and feed it into the parser.

As there are no longer any temporary files to clean up, KmlDocument is
no longer needed, GeoDataDocument is sufficient.

This also avoids an error in the thumbnailer, which fails when it
tries to geocast the KmlDocument to a GeoDataDocument.

BUG: 428437
CCBUG: 420089
CCBUG: 416925