Avoid the need to set Qt::WA_DeleteOnClose on widgets using Marble
I've noticed that when using MarbleWidget in my app, closing the top
level widget will no longer quit the QApplication. This is because the
PopupLayer is still alive and has Qt::WA_QuitOnClose set, but is never
closed.

When unsetting it, MarbleWidget can be used in a window without having
to set Qt::WA_DeleteOnClose. This allows the window to be created on the
stack - also how it's done in Qt's examples:
https://doc.qt.io/qt-5/qtwidgets-mainwindows-application-example.html