Fix building Marble for distributions
LICENSE.txt is a symlink, but it gets installed as a symlink (not as a
file, as INSTALL copies symlinks as symlinks) to the
license in LICENSES/, which is not installed. Therefore build systems
that watch for files that are not installed will notice a symlink
pointing to a location in the source and not on the target path and will
error out.

Having INSTALL handle symlinks is tricky, so the best solution is to
just install the actual file in LICENSES and rename it.