Protect nodes adjacent to synthetic clip nodes during node reduction
This is to avoid ending up with a synthetic-only line that client-side
geometry assembly might drop entirely, and thus distorting the original
shape.

Example: An 'L'-shaped line with the knee of the L barely extending into
a separate tile. Previously we might have dropped the node at the knee
there and only retained the two synthetic clip nodes at the tile border.
When reassembling this with the two legs of the L that segment can get
removed, resulting in a diagonal line between the two ends of the L.

With this change we retain the knee node, no matter how close it might
be to the clip node, and thus geometry reassembly managed to restore the
proper shape.