Don't emit ways again if they are already part of a multi-polygon
In case one of the individual rings of a multi-polygon has tags, it can
exist as its own placemark object next to the multi-polyon one as well.
Due to that we duplicated those ways in the output tiles. That not only
needlessly increases them in size, it also poses a significant challenge
for geometry reassembly.