Restore former GPX rendering behavior.
Previous versions of Marble rendered GPX tracks using a thick red line
and then recentered and zoomed the map so that the newly-loaded track
occupied most of the display.  At some point, this behavior got lost;
for some time now, Marble has been rendering GPX tracks as a
difficult-to-see 1-pixel-wide white line and not recentering or zooming
the map, leading some users to the conclusion that the software wasn't
loading the track at all.  See the following bug reports:

  https://bugs.kde.org/show_bug.cgi?id=419998
  https://bugs.kde.org/show_bug.cgi?id=411669
  https://bugs.kde.org/show_bug.cgi?id=370263
  https://bugs.kde.org/show_bug.cgi?id=402153

This patch restores the old rendering behavior.